Refactor handler_test.gp
